+++ samples/react-native/babel.config.js
@@ -1,5 +1,3 @@
-const componentAnnotatePlugin = require('@sentry/babel-plugin-component-annotate');
-
 module.exports = {
   presets: ['module:@react-native/babel-preset'],
   plugins: [
@@ -12,6 +10,5 @@ module.exports = {
       },
     ],
     'react-native-reanimated/plugin',
-    componentAnnotatePlugin,
   ],
 };