+++ samples/react-native/src/Screens/TrackerScreen.tsx
@@ -73,9 +73,7 @@ const TrackerScreen = () => {
   return (
     <View style={styles.screen}>
       <Sentry.TimeToInitialDisplay record />
-      <View style={styles.titleContainer}>
-        <Text style={styles.title}>Global COVID19 Cases</Text>
-      </View>
+      <TrackerTitle />
       <View style={styles.card}>
         {cases ? (
           <>
@@ -110,6 +108,12 @@ const TrackerScreen = () => {
     </View>
   );
 };
+
+const TrackerTitle = () => (
+  <View style={styles.titleContainer}>
+    <Text style={styles.title}>Global COVID19 Cases</Text>
+  </View>
+);
 
 export default Sentry.withProfiler(TrackerScreen);