+++ test/tools/fixtures/mockBabelTransformer.js
@@ -1,0 +1,4 @@
+module.exports = {
+  transform: jest.fn(),
+  getCacheKey: jest.fn(),
+};